requirements.txt
nltk==3.6.5
pandas==1.3.4
streamlit==1.2.0
plotly==5.3.1
python-dotenv==0.19.2
pymongo==3.12.1
transformers==4.12.5
torch==1.10.0
scikit-learn==1.0.1
textblob==0.15.3
